% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmgt_compact.R
\name{dmgt_compact}
\alias{dmgt_compact}
\title{Compact data frame list into a single element per table}
\usage{
dmgt_compact(
  .data,
  .data_year = NULL,
  .agency = NULL,
  .study = NULL,
  .save_raw_csv = FALSE
)
}
\arguments{
\item{.data}{A list containing multiple data frames}

\item{.data_year}{Year the data was collected}

\item{.agency}{The agency collecting the data}

\item{.study}{The study the field work was conducted under}

\item{.save_raw_csv}{Do you want to save the raw data to .csv}
}
\value{
a concise list of data frames
}
\description{
Compact data frame list into a single element per table
}
