% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmgt_compact_2.R
\name{dmgt_compact_2}
\alias{dmgt_compact_2}
\title{Title}
\usage{
dmgt_compact_2(.data)
}
\arguments{
\item{.data}{A list of data frames extracted from .dbf files}
}
\value{
A list conisiting of only named data frames
}
\description{
Title
}
