% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmgt_transform_csu.R
\name{dmgt_transform_csu}
\alias{dmgt_transform_csu}
\title{Transform variables in CSU/CPW nnf removal applications}
\usage{
dmgt_transform_csu(.data)
}
\arguments{
\item{.data}{A list of data frames}
}
\value{
A list of data frames
}
\description{
Transform variables in CSU/CPW nnf removal applications
}
