% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_fish.R
\name{flag_fish}
\alias{flag_fish}
\title{Add QC flags to standard fish data frame}
\usage{
flag_fish(.data)
}
\arguments{
\item{.data}{A data frame standardized to the fish spec}
}
\value{
A data frame
}
\description{
Add QC flags to standard fish data frame
}
