% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_site.R
\name{flag_site}
\alias{flag_site}
\title{Add QC flags to standard site data frame}
\usage{
flag_site(.data)
}
\arguments{
\item{.data}{A data frame standardized to the site spec}
}
\value{
A data frame
}
\description{
Add QC flags to standard site data frame
}
