% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_water.R
\name{flag_water}
\alias{flag_water}
\title{Add QC flags to standard water data frame}
\usage{
flag_water(.data)
}
\arguments{
\item{.data}{A data frame standardized to the water spec}
}
\value{
A data frame
}
\description{
Add QC flags to standard water data frame
}
