% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dbf.R
\name{read_dbf}
\alias{read_dbf}
\title{Reads multiple dbf files}
\usage{
read_dbf(.file_path_in, .file_pattern)
}
\arguments{
\item{.file_path_in}{The file path to the target directory}

\item{.file_pattern}{The pattern used to target the files}
}
\value{
A list of data frames
}
\description{
Reads multiple dbf files
}
