% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datatype.R
\name{test_datatype}
\alias{test_datatype}
\title{Check data types and column names from DataPlus raw data frame}
\usage{
test_datatype(.data, .template)
}
\arguments{
\item{.data}{The raw DataPlus data frame}

\item{.template}{The applicable data template}
}
\value{
A data frame and message
}
\description{
Check data types and column names from DataPlus raw data frame
}
