% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_data_standard}
\alias{tmplt_data_standard}
\title{Data typed standard data structure templates}
\format{
A list containing 3 data frame templates
}
\usage{
tmplt_data_standard
}
\description{
This set of templates reflects the current data standard
}
\keyword{datasets}
