% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_DataPlus}
\alias{tmplt_DataPlus}
\title{Data typed DataPlus structure templates}
\format{
A list containing 6 data frame templates
}
\usage{
tmplt_DataPlus
}
\description{
This set of templates reflects the data structure of the DataPlus app
}
\keyword{datasets}
